**Refactor CodeGen and MicroBuilder: Improve const correctness and remove redundant return statements**
- Converted local variables to `constexpr` and `auto` where possible for improved clarity and consistency.
- Aligned variable declarations for better readability in multiple modules, including `CodeGen.Intrinsic`, `CodeGen.Cast`, and `CodeGen.Constant`.
- Removed unnecessary `return` statements in `MicroBuilder`, `X64Encoder`, and `ConstantLower` for simpler control flow.
- Streamlined alignment logic for slice and entry offsets in `CodeGen.Function`.

+++ src/Backend/Micro/Passes/Pass.ConstantPropagation.Helpers.cpp
@@ -1,10 +1,10 @@
 #include "pch.h"
-#include "Main/TaskContext.h"
-#include "Compiler/Sema/Constant/ConstantManager.h"
 #include "Backend/ABI/CallConv.h"
 #include "Backend/Micro/MicroPassContext.h"
 #include "Backend/Micro/MicroPassHelpers.h"
 #include "Backend/Micro/Passes/Pass.ConstantPropagation.h"
+#include "Compiler/Sema/Constant/ConstantManager.h"
+#include "Main/TaskContext.h"
 
 SWC_BEGIN_NAMESPACE();
 
@@ -199,20 +199,20 @@ bool MicroConstantPropagationPass::constantPointerRangeHasRelocation(const Known
     if (!numBytes)
         return false;
 
-    uint32_t       shardIndex = 0;
-    const Ref      baseOffset = context_->taskContext->cstMgr().findDataSegmentRef(shardIndex, reinterpret_cast<const void*>(constantPointer.pointer));
+    uint32_t  shardIndex = 0;
+    const Ref baseOffset = context_->taskContext->cstMgr().findDataSegmentRef(shardIndex, reinterpret_cast<const void*>(constantPointer.pointer));
     if (baseOffset == INVALID_REF)
         return true;
 
-    const DataSegment& segment     = context_->taskContext->cstMgr().shardDataSegment(shardIndex);
-    const uint64_t     rangeBegin  = static_cast<uint64_t>(baseOffset) + constantPointer.offset;
-    const uint64_t     rangeEnd    = rangeBegin + numBytes;
-    constexpr uint64_t relocSize   = sizeof(uint64_t);
+    const DataSegment& segment    = context_->taskContext->cstMgr().shardDataSegment(shardIndex);
+    const uint64_t     rangeBegin = static_cast<uint64_t>(baseOffset) + constantPointer.offset;
+    const uint64_t     rangeEnd   = rangeBegin + numBytes;
+    constexpr uint64_t relocSize  = sizeof(uint64_t);
 
     for (const auto& relocation : segment.relocations())
     {
         const uint64_t relocBegin = relocation.offset;
-        if (MicroPassHelpers::rangesOverlap(rangeBegin, numBytes, relocBegin, static_cast<uint32_t>(relocSize)))
+        if (MicroPassHelpers::rangesOverlap(rangeBegin, numBytes, relocBegin, relocSize))
             return true;
         if (relocBegin >= rangeEnd)
             break;
